UpdateReturns/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.38.0
numpy==1.20.3
pandas==1.2.4